fix(convex): break storage/r2 import cycle via leaf r2Keys module
